feat(runtime): raise default max_iterations to 80
Align with OpenClaw's default maxTurns of 80 to give complex
multi-tool tasks enough headroom. Also fixes two pre-existing
clippy warnings (unnecessary_cast in orchestrator, to_string_in_format_args
in slack runner) that were blocking lint.

+++ crates/runtime/src/orchestrator.rs
@@ -499,10 +499,8 @@ impl Orchestrator {
 
                             let obs = match exec_result {
                                 Ok(output) => {
-                                    otel_span.set_attribute(KeyValue::new(
-                                        "duration_ms",
-                                        duration_ms as i64,
-                                    ));
+                                    otel_span
+                                        .set_attribute(KeyValue::new("duration_ms", duration_ms));
                                     otel_span.set_attribute(KeyValue::new("tool_status", "ok"));
                                     otel_span.set_attribute(KeyValue::new(
                                         "tool_observation",
@@ -514,10 +512,8 @@ impl Orchestrator {
                                 Err(err) => {
                                     warn!(tool = %name, %err, "Extension tool execution failed");
                                     let msg = err.to_string();
-                                    otel_span.set_attribute(KeyValue::new(
-                                        "duration_ms",
-                                        duration_ms as i64,
-                                    ));
+                                    otel_span
+                                        .set_attribute(KeyValue::new("duration_ms", duration_ms));
                                     otel_span.set_attribute(KeyValue::new("tool_status", "error"));
                                     otel_span
                                         .set_attribute(KeyValue::new("tool_error", msg.clone()));
